Move logic wrapper to new file.
The StatsCommercialUpgradeSlider serves as a logic wrapper around the TierUpgradeSlider and handles context specific to the commerical upgrade flow.
