+++ client/my-sites/stats/stats-purchase/stats-commercial-upgrade-slider/styles.scss
@@ -1,0 +1,5 @@
+// Styles specific to layout on the Commercial Upgrade page.
+
+.stats-commercial-upgrade-slider {
+	margin: 42px 0;
+}
